fix again

+++ android/lib/src/main/java/com/airbnb/android/react/maps/AirMapManager.java
@@ -89,12 +89,12 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     }
 
     @ReactProp(name = "maxZoom")
-    public void setMaxZoom(AirMapView view, Double maxZoom) {
+    public void setMaxZoom(AirMapView view, Float maxZoom) {
         view.setMaxZoom(maxZoom);
     }
 
     @ReactProp(name = "minZoom")
-    public void setMinZoom(AirMapView view, Double minZoom) {
+    public void setMinZoom(AirMapView view, Float minZoom) {
         view.setMinZoom(minZoom);
     }
